# Optimize RLE+ Validation by specialized decoder