Benchmark:
```
Count/0/basic-8              193ns ± 7%   122ns ±16%  -36.72%  (p=0.008 n=5+5)
Count/1/basic-8              211ns ± 8%   134ns ± 3%  -36.28%  (p=0.008 n=5+5)
Count/1/modified-8           858ns ± 5%   772ns ± 6%  -10.00%  (p=0.008 n=5+5)
Count/10/basic-8             356ns ± 5%   242ns ± 3%  -31.97%  (p=0.008 n=5+5)
Count/10/modified-8         1.42µs ± 3%  1.21µs ± 4%  -14.53%  (p=0.008 n=5+5)
Count/1000/basic-8          23.3µs ± 5%  15.0µs ± 0%  -35.52%  (p=0.008 n=5+5)
Count/1000/modified-8       45.8µs ± 2%  37.1µs ± 0%  -19.02%  (p=0.016 n=5+4)
Count/1000000/basic-8       24.7ms ± 2%  17.3ms ± 2%  -29.94%  (p=0.008 n=5+5)
Count/1000000/modified-8    44.1ms ± 0%  36.9ms ± 0%  -16.42%  (p=0.016 n=5+4)
IsEmpty/0/basic-8            191ns ± 9%   127ns ± 7%  -33.72%  (p=0.008 n=5+5)
IsEmpty/1/basic-8            206ns ± 4%   137ns ± 5%  -33.53%  (p=0.008 n=5+5)
IsEmpty/1/modified-8         889ns ± 5%   830ns ± 9%     ~     (p=0.095 n=5+5)
IsEmpty/10/basic-8           298ns ± 3%   184ns ± 6%  -38.32%  (p=0.008 n=5+5)
IsEmpty/10/modified-8       1.17µs ±11%  1.05µs ± 5%  -10.55%  (p=0.032 n=5+5)
IsEmpty/1000/basic-8        11.2µs ± 4%   4.3µs ± 2%  -61.51%  (p=0.008 n=5+5)
IsEmpty/1000/modified-8     12.0µs ± 0%   5.3µs ± 1%  -55.90%  (p=0.016 n=4+5)
IsEmpty/1000000/basic-8     12.1ms ± 2%   4.8ms ± 2%  -59.98%  (p=0.008 n=5+5)
IsEmpty/1000000/modified-8  11.9ms ± 3%   4.7ms ± 0%  -60.21%  (p=0.016 n=5+4)
BigDecodeEncode-8            756µs ± 3%   600µs ± 2%  -20.54%  (p=0.008 n=5+5)
BigValidate-8                278µs ± 3%   130µs ± 3%  -53.29%  (p=0.008 n=5+5)
```